refactor: migrate announcement feature to Ktor
This commit migrates the announcement feature from Retrofit to Ktor for network requests.

Key changes include:
- Renamed `AnnouncementRepositoryImpl` from `domain` to `data` package.
- Renamed `AnnouncementResponse` and `Announcement` DTOs to a `dto` subpackage.
- Updated `AnnouncementRepository` and `AnnouncementRepositoryImpl` to use Ktor client and its `Result` type for network operations.
- Modified `AnnouncementViewModel` to handle Ktor's `Result` and use `UiText` for notification messages.
- Adjusted `AnnouncementScreen` to consume `UiText` from the ViewModel.
- Updated import paths for DTOs in `AnnouncementUi`, `AnnouncementState`, and `AnnouncementEvent`.
