+++ app/src/main/java/com/mycheva/app/announcement/domain/AnnouncementRepository.kt
@@ -1,14 +1,16 @@
 package com.mycheva.app.announcement.domain
 
-import com.mycheva.app.announcement.data.AnnouncementsItem
+import com.mycheva.app.announcement.data.dto.AnnouncementResponse
 import com.mycheva.app.core.database.BookmarkEntity
+import com.mycheva.app.core.network.utils.DataError
+import com.mycheva.app.core.network.utils.Result
 import kotlinx.coroutines.flow.Flow
 
 interface AnnouncementRepository {
 
     fun getToken(): Flow<String>
 
-    suspend fun getAnnouncements(token: String): Result<List<AnnouncementsItem>>
+    suspend fun getAnnouncements(token: String): Result<AnnouncementResponse, DataError.Remote>
 
     suspend fun addBookmark(bookmark: BookmarkEntity): Boolean
