+++ app/src/main/java/com/mycheva/app/announcement/presentation/AnnouncementEvent.kt
@@ -1,6 +1,6 @@
 package com.mycheva.app.announcement.presentation
 
-import com.mycheva.app.announcement.data.AnnouncementsItem
+import com.mycheva.app.announcement.data.dto.AnnouncementsItem
 
 sealed class AnnouncementEvent {
     data class OnAddBookmark(val announcement: AnnouncementsItem): AnnouncementEvent()
