+++ app/src/main/java/com/mycheva/app/announcement/presentation/AnnouncementState.kt
@@ -1,11 +1,12 @@
 package com.mycheva.app.announcement.presentation
 
-import com.mycheva.app.announcement.data.AnnouncementsItem
+import com.mycheva.app.announcement.data.dto.AnnouncementsItem
+import com.mycheva.app.core.ui.utils.UiText
 
 data class AnnouncementState(
     val announcements: List<AnnouncementsItem> = emptyList(),
     val isLoading: Boolean = false,
-    val notificationMessage: String = "",
+    val notificationMessage: UiText = UiText.DynamicString(""),
     val isRequestError: Boolean = false,
     val token: String = "",
 )
